pyproject.toml
[tool.poetry]
name = "gallery-api"
version = "0.1.0"
description = ""
authors = ['Kimiyori']
readme = "README.md"
packages = [{include = "gallery_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
environs = "^9.5.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^3.7.0"
dependency-injector = "^4.40.0"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.4"
pillow = "^9.3.0"

[tool.poetry.group.tests.dependencies]
pytest = "^7.2.0"
pytest-django = "^4.5.2"
faker = "^15.3.3"
[tool.poetry.group.linters.dependencies]
black = "^22.10.0"
flake8 = "^6.0.0"
pylint = "^2.15.6"
mypy = "0.982"
django-stubs = {extras = ["compatible-mypy"], version = "^1.13.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
